--- a/src/cmd/cmd.zig
+++ b/src/cmd/cmd.zig
@@ -376,6 +376,7 @@ fn gossip() !void {
         &.{},
     );
     defer gossip_service.deinit();
+
     try runGossipWithConfigValues(&gossip_service);
 }
 
@@ -571,8 +572,16 @@ fn initRepair(
 }
 
 fn runGossipWithConfigValues(gossip_service: *GossipService) !void {
+    // TODO: use better allocator, unless GPA becomes more performant.
+    var gp_message_allocator: std.heap.GeneralPurposeAllocator(.{}) = .{};
+    defer _ = gp_message_allocator.deinit();
+
     const gossip_config = config.current.gossip;
-    return gossip_service.run(gossip_config.spy_node, gossip_config.dump);
+    return gossip_service.run(.{
+        .message_allocator = gp_message_allocator.allocator(),
+        .spy_node = gossip_config.spy_node,
+        .dump = gossip_config.dump,
+    });
 }
 
 /// determine our shred version and ip. in the solana-labs client, the shred version