Starts work towards #119 & #120.
More details in the commits.