@InKryption 2024-06-03
I conducted many more runs than presented here, but those presented are representative of the general trends I observed, where performance is either equivalent or minimally better, excluding outliers where variance exceeded more than 100 in either (which occurred twice with the old code, and once with the new code).

Side note: brief testing indicated that ReleaseFast demonstrates near identical trends.

Before socket_utils changes:
```
❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils  
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1739       1755         43       1747



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1737       1755         56       1745



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1743       1751         12       1747



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1739       1742          2       1740



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1740       1747          8       1743
```

After socket_utils changes:
```
❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1719       1735         44       1727



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1729       1734          4       1732



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1730       1733          1       1731



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1728       1738         17       1732



❯ zig build -Doptimize=ReleaseSafe benchmark -- socket_utils
filtering benchmarks with prefix: socket_utils

Benchmark                      Iterations    Min(ms)    Max(ms)   Variance   Mean(ms)
-------------------------------------------------------------------------------------
benchmarkReadSocket(100k_msgs)          3       1721       1738         51       1728
```